# docs/api/openapi.yaml
openapi: 3.1.2
info:
  title: E2B Billing API
  version: 0.4.0
  description: |
    The usage endpoint validates and atomically commits measurements to the
    PostgreSQL inbox before returning 202. Identical event retries succeed
    without changing stored receipt or accounting state; changed content
    conflicts. Acceptance confirms durable receipt, while accounting remains
    asynchronous in the standalone worker. Unsupported customers, schemas,
    metrics, or crossing intervals are retained with a processing error.
servers:
  - url: /api
    description: Same-origin API proxy provided by the local Scalar service
...
paths:
  /usage/batches:
    post:
      operationId: postUsageBatches
      tags: [Usage]
      summary: Commit a usage batch to the inbox
      x-implementation-status: implemented
      description: |
        Parse one uncompressed UTF-8 JSON document, validate its envelope and
        every measurement, and commit the whole batch in one inbox transaction.
        Return 202 only after a successful synchronous PostgreSQL commit.

        The body is limited to 1048576 bytes and 1–1000 events. Member names are
        case-sensitive; unknown and duplicate members, malformed UTF-8, and
        unpaired UTF-16 escapes are rejected. Identifiers and an optional
        batch_id are limited to 256 UTF-8 bytes each. Required event fields must
        be explicitly supplied and non-null, including units when its value is
        zero. Version and units use integer tokens without fractional or
        exponent notation; decoding outside int32/int64 ranges returns 400.

        Timestamps use RFC 3339 with an explicit offset, valid calendar values,
        UTC years 1000–9999, and at most six fractional digits. Interval ordering
        compares instants. Identifiers are preserved; instants normalize to UTC.
        Positive schema versions and nonblank metrics are structurally accepted;
        customer existence, schema version 1, and historical prices are checked
        asynchronously by the accounting worker. A receipt must fit one UTC
        month and one applicable price version; unsupported input is retained
        with processing_error and contributes no financial effect.

        Identify measurements by (source, event_id) and accept a whole batch
        atomically, including repeated identities within the incoming batch.
        Identical content succeeds across retries and batch boundaries without
        another insert or a changed received_at. Changed measurement content
        for an existing identity returns 409 and rolls back new events in the
        batch. Billing supplies the first received_at from its UTC clock.
        Compare schema_version, customer_id, sandbox_id, metric, period_start,
        period_end, and units as domain values; timestamp offsets representing
        the same instant are identical. Reordered events and different batch
        boundaries do not change event identity. Batch identifiers are optional
        producer metadata, are not stored, and do not participate in comparison.
        Accounting remains asynchronous and is not implied by acceptance.

        Admission is bounded per API instance (32 concurrent batches by
        default). When full, return 503 before application body parsing;
        retain the original events for retry. Database work, including waiting
        for a pool connection, has a configurable deadline (10 seconds by
        default), also honoring request cancellation. Admission exhaustion,
        unavailable storage, or a failed commit returns 503 with
        Retry-After: 1. A 503 or a lost response can leave the commit outcome
        unknown: retain events and retry the same identities and content after
        backoff. Identical retries are safe even after successful accounting.
        A 409 requires investigation; unchanged retries cannot resolve it.
      requestBody:
        required: true
        description: A nonempty batch of explicit measurement values.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsageBatch'
            examples:
              acmeUsage:
                summary: First Acme measurement from the assignment
                value:
                  batch_id: acme-batch-000001
                  events:
                    - source: platform-simulator
                      event_id: acme-cpu-000001
                      schema_version: 1
                      customer_id: acme
                      sandbox_id: acme-sandbox-001
                      metric: cpu_seconds
                      period_start: '2026-10-10T12:00:00Z'
                      period_end: '2026-10-10T13:00:00Z'
                      units: 100000000
      responses:
        '202':
          description: |
            Every new measurement is durably committed; existing identical
            measurements are preserved, including their original received_at,
            processed_at, and processing_error. No partial batch is accepted.
            Accounting completion is not implied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedResponse'
              example:
                status: accepted
        '400':
          description: |
            Malformed or ambiguous JSON, incorrect JSON types, or numeric
            tokens outside the representable int32/int64 ranges. No batch is
            accepted. Correct the request before retrying.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: invalid_json
                  message: The object contains a duplicate field.
                  field: events[0]
        '413':
          description: The body exceeds 1048576 bytes. Split the batch before retrying.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: request_too_large
                  message: The request body must not exceed 1048576 bytes.
        '415':
          description: |
            Content-Type must be application/json with optional UTF-8 charset.
            Content-Encoding must be omitted or identity; compression is not
            supported. Correct the headers and encoding before retrying.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: unsupported_media_type
                  message: Use application/json with UTF-8 encoding.
        '422':
          description: |
            A required field is absent/null or a value violates the batch or
            measurement contract. No partial batch is accepted. Correct the
            reported value before retrying; zero units must be explicit.
            An instant whose UTC year is outside 1000–9999 is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidInterval:
                  value:
                    error:
                      code: invalid_batch
                      message: must be later than period_start
                      field: events[0].period_end
                unsupportedUTCYear:
                  value:
                    error:
                      code: invalid_batch
                      message: must have a UTC year between 1000 and 9999
                      field: events[0].period_start
        '409':
          description: |
            Different measurement content for one (source, event_id), either
            already stored or repeated inside this request. Original rows are
            preserved and every new row in the batch is rolled back. The
            response identifies the conflicting key, without exposing its
            stored content. Retain the input for investigation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictResponse'
              example:
                error:
                  code: event_conflict
                  message: The event identity already has different measurement content.
                  source: platform-simulator
                  event_id: acme-cpu-000001
        '503':
          description: |
            The API's concurrent batch budget is exhausted, or inbox storage is
            unavailable, timed out, canceled, or failed during commit, including
            a missing schema before explicit migrations. Admission rejection
            happens before application body parsing and writes no events. The
            commit outcome may be unknown. Retain events and retry unchanged
            identities/content after backoff; storage details are not exposed.
          headers:
            Retry-After:
              description: Minimum delay in seconds before retrying; value is 1.
              schema:
                type: string
                const: '1'
              example: '1'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: inbox_unavailable
                  message: Inbox storage is unavailable. Retry the same event identities and content.
...
components:
  schemas:
    Identifier:
      type: string
      minLength: 1
      maxLength: 256
      description: |
        Preserved text containing at least one non-whitespace character, valid
        Unicode, and no NUL. The actual maximum is 256 UTF-8 bytes, including
        leading/trailing whitespace; maxLength also bounds its character count.
      allOf:
        - pattern: '^[^\u0000]*$'
        - pattern: '[^\u0009-\u000d\u0020\u0085\u00a0\u1680\u2000-\u200a\u2028\u2029\u202f\u205f\u3000]'
    UsageBatch:
      type: object
      additionalProperties: false
      required: [events]
      properties:
        batch_id:
          $ref: '#/components/schemas/Identifier'
          description: Optional producer metadata; not stored and not used for deduplication.
        events:
          type: array
          minItems: 1
          maxItems: 1000
          items:
            $ref: '#/components/schemas/UsageEvent'
    ConsumptionTime:
      type: string
      format: date-time
      pattern: '^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}(\.[0-9]{1,6})?(Z|[+-]([01][0-9]|2[0-3]):[0-5][0-9])$'
      description: |
        Valid RFC 3339 instant with an explicit offset and at most six
        fractional digits. The corresponding UTC year must be 1000–9999.
        The minimum instant is 1000-01-01T00:00:00Z. Bounds apply after UTC
        conversion: a local year of 1000 can be invalid if its offset moves
        the instant into 0999, while a local year of 0999 can be valid if
        its offset moves the instant into 1000.
    UsageEvent:
      type: object
      additionalProperties: false
      description: |
        Measured increment over [period_start, period_end). The end must follow
        the start as an instant. Fields must be explicit and non-null. Units
        are neither a cumulative counter nor a monetary amount.
      required: [source, event_id, schema_version, customer_id, sandbox_id, metric, period_start, period_end, units]
      properties:
        source:
          $ref: '#/components/schemas/Identifier'
          description: Stable producer namespace preserved across restarts and retries.
        event_id:
          $ref: '#/components/schemas/Identifier'
          description: Identity unique within source; preserved with content on retries.
        schema_version:
          type: integer
          format: int32
          minimum: 1
          maximum: 2147483647
          description: Positive contract version encoded as an integer token.
        customer_id:
          $ref: '#/components/schemas/Identifier'
        sandbox_id:
          $ref: '#/components/schemas/Identifier'
        metric:
          $ref: '#/components/schemas/Identifier'
          description: Resource identifier; accounting registries are future work.
        period_start:
          $ref: '#/components/schemas/ConsumptionTime'
          description: Inclusive start of consumption.
        period_end:
          $ref: '#/components/schemas/ConsumptionTime'
          description: Exclusive end; must be later than period_start.
        units:
          type: integer
          format: int64
          minimum: 0
          maximum: 9223372036854775807
          description: Explicit increment encoded as an integer token; zero is valid.
    AcceptedResponse:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          type: string
          enum: [accepted]
          description: Durable inbox receipt; accounting is asynchronous.
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
              enum: [invalid_json, invalid_batch, request_too_large, unsupported_media_type, event_conflict, inbox_unavailable]
            message:
              type: string
              description: Human-readable description; clients should use code for classification.
            field:
              type: string
              description: Optional location, such as events[0].units.
            source:
              $ref: '#/components/schemas/Identifier'
              description: Producer namespace of a conflicting event; present on 409.
            event_id:
              $ref: '#/components/schemas/Identifier'
              description: Conflicting identity within source; present on 409.
    ConflictResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - properties:
            error:
              required: [source, event_id]
              properties:
                code:
                  const: event_conflict
...